Slightly modified the submission handler
Form submission was behaving wonky, found that the length of the fieldsets array was actually 1 longer than expected as the submit button was also wrapped in an HTML fieldset.

Hardcoded the array length values and submit handler is working as expected. Also slightly refactored the validation code for the form, will only validate the form if the submission event is triggered on the final fieldset, then it checks if all field requirements are valid and submits the form, otherwise it throws an error to the console